--- src/shared/components/button/index.ts
@@ -1,3 +1,0 @@
-import Button from "./Button";
-
-export default Button;
