+++ src/shared/components/divider.tsx
