+++ src/shared/components/image-detail-overlay.tsx
