+++ src/shared/components/index.ts
@@ -1,18 +1,18 @@
-export { default as Button } from "./button";
+export { default as Authentication } from "./authentication";
 export { default as Blockquote } from "./blockquote";
+export { default as Button } from "./button";
+export { default as Checkbox } from "./checkbox";
 export { default as Divider } from "./divider";
+export { default as Dropdown } from "./dropdown";
+export { default as ErrorMessage } from "./error-message";
 export { default as If } from "./if";
-export { default as ToastBase } from "./toast";
+export { default as ImageDetailOverlay } from "./image-detail-overlay";
 export { default as Modal } from "./modal";
 export { default as Portal } from "./portal";
-export { default as Table } from "./table";
 export { default as SelectBox } from "./select-box";
-export { default as Switch } from "./switch";
-export { default as Checkbox } from "./checkbox";
-export { default as Dropdown } from "./dropdown";
+export { default as Skeleton } from "./skeleton";
 export { default as Stepper } from "./stepper";
-export { default as Authentication } from "./authentication";
+export { default as Switch } from "./switch";
+export { default as Table } from "./table";
 export { default as TextArea } from "./textarea";
-export { default as ErrorMessage } from "./error-message";
-export { default as ImageDetailOverlay } from "./image-detail-overlay";
-export * from "./skeleton";
+export { default as ToastBase } from "./toast";
