--- src/shared/components/skeleton/image.tsx
@@ -1,10 +1,0 @@
-const Image = () => {
-  return (
-    <div
-      role="status"
-      className={`flex items-center justify-center w-full h-96 bg-gray-300 rounded-lg animate-pulse`}
-    />
-  );
-};
-
-export default Image;
