--- src/shared/components/skeleton/index.tsx
@@ -1,3 +1,0 @@
-export { default as Skeleton } from "./default";
-export { default as ImageSkeleton } from "./image";
-export { default as TextSkeleton } from "./text";
