+++ src/shared/hooks/use-overlay.tsx
@@ -1,6 +1,5 @@
-import { useContext, useMemo, useState } from "react";
+import { useContext, useEffect, useMemo, useState } from "react";
 
-import { OverlayController } from "../components/overlay-controller";
 import { OverlayContext } from "../contexts";
 
 let elementId = 1;
@@ -8,7 +7,28 @@ let elementId = 1;
 type CreateOverlayElement = (props: {
   isOpen: boolean;
   exit: () => void;
-}) => JSX.Element;
+}) => JSX.Element | null;
+
+interface Props {
+  overlayElement: CreateOverlayElement;
+  onExit: () => void;
+}
+
+const OverlayController = ({
+  overlayElement: OverlayElement,
+  onExit,
+}: Props) => {
+  const [isOpenOverlay, setIsOpenOverlay] = useState(false);
+
+  useEffect(() => {
+    // NOTE: requestAnimationFrame이 없으면 가끔 Open 애니메이션이 실행되지 않는다.
+    requestAnimationFrame(() => {
+      setIsOpenOverlay(true);
+    });
+  }, []);
+
+  return <OverlayElement isOpen={isOpenOverlay} exit={onExit} />;
+};
 
 const useOverlay = () => {
   const context = useContext(OverlayContext);
